+++ src/main/java/org/entur/enlil/configuration/EnlilConfiguration.java
@@ -1,9 +1,15 @@
 package org.entur.enlil.configuration;
 
+import java.time.Clock;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
+import org.springframework.scheduling.annotation.EnableScheduling;
 
 @Configuration
+@EnableScheduling
 public class EnlilConfiguration {
 
   @Value("${enlil.siri.default.producerRef:ENT}")
@@ -11,5 +17,11 @@ public class EnlilConfiguration {
 
   public String getProducerRef() {
     return producerRef;
+  }
+
+  @Bean
+  @ConditionalOnMissingBean
+  Clock clock() {
+    return Clock.systemDefaultZone();
   }
 }
