+++ src/main/java/org/entur/enlil/graphql/MutationController.java
@@ -1,7 +1,7 @@
-package org.entur.enlil.siri.graphql;
+package org.entur.enlil.graphql;
 
-import org.entur.enlil.siri.repository.SituationElementRepository;
-import org.entur.enlil.siri.repository.firestore.entity.PtSituationElementEntity;
+import org.entur.enlil.model.PtSituationElementEntity;
+import org.entur.enlil.repository.SituationElementRepository;
 import org.springframework.graphql.data.method.annotation.Argument;
 import org.springframework.graphql.data.method.annotation.MutationMapping;
 import org.springframework.graphql.data.method.annotation.SchemaMapping;
