+++ src/main/java/org/entur/enlil/repository/EstimatedVehicleJourneyRepository.java
@@ -1,0 +1,8 @@
+package org.entur.enlil.repository;
+
+import java.util.stream.Stream;
+import org.entur.enlil.model.EstimatedVehicleJourneyEntity;
+
+public interface EstimatedVehicleJourneyRepository {
+  Stream<EstimatedVehicleJourneyEntity> getAllEstimatedVehicleJourneys();
+}
