+++ src/main/java/org/entur/enlil/repository/firestore/FirestoreSituationElementRepository.java
@@ -1,4 +1,4 @@
-package org.entur.enlil.siri.repository.firestore;
+package org.entur.enlil.repository.firestore;
 
 import com.google.cloud.firestore.Filter;
 import com.google.cloud.firestore.Firestore;
@@ -10,14 +10,12 @@ import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 import java.util.stream.Stream;
+import org.entur.enlil.model.PtSituationElementEntity;
+import org.entur.enlil.repository.SituationElementRepository;
 import org.entur.enlil.siri.helpers.DateMapper;
-import org.entur.enlil.siri.repository.SituationElementRepository;
-import org.entur.enlil.siri.repository.firestore.entity.PtSituationElementEntity;
-import org.entur.enlil.siri.repository.firestore.mapper.SituationElementEntityToSiriMapper;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Repository;
-import uk.org.siri.siri21.PtSituationElement;
 
 @Repository
 public class FirestoreSituationElementRepository implements SituationElementRepository {
@@ -33,10 +31,8 @@ public class FirestoreSituationElementRepository implements SituationElementRepo
     this.clock = clock;
   }
 
-  public Stream<PtSituationElement> getAllSituationElements() {
-    return Stream
-      .concat(getOpenMessages(), getClosedValidMessages())
-      .map(SituationElementEntityToSiriMapper::mapToPtSituationElement);
+  public Stream<PtSituationElementEntity> getAllSituationElements() {
+    return Stream.concat(getOpenMessages(), getClosedValidMessages());
   }
 
   private Stream<PtSituationElementEntity> getOpenMessages() {
