+++ src/test/java/org/entur/enlil/EnlilApplicationIntegrationTests.java
@@ -15,9 +15,9 @@ import java.time.temporal.ChronoUnit;
 import java.util.List;
 import java.util.concurrent.ExecutionException;
 import org.entur.enlil.configuration.MockedClockConfiguration;
-import org.entur.enlil.siri.repository.firestore.entity.EstimatedVehicleJourneyEntity;
-import org.entur.enlil.siri.repository.firestore.entity.FramedVehicleJourneyRef;
-import org.entur.enlil.siri.repository.firestore.entity.PtSituationElementEntity;
+import org.entur.enlil.model.EstimatedVehicleJourneyEntity;
+import org.entur.enlil.model.FramedVehicleJourneyRef;
+import org.entur.enlil.model.PtSituationElementEntity;
 import org.jetbrains.annotations.NotNull;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
